requirements.txt
bcrypt==4.0.1
beautifulsoup4==4.12.1
bleach==6.0.0
blinker==1.6.2
boolean.py==4.0
cffi==1.16.0
click==8.1.7
cryptography==43.0.1
Flask==2.3.2
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.0.3
Flask-Session>=0.3.2,<0.6
Flask-WTF==1.2.0
greenlet==3.0.3
identity>=0.5.1,<0.6
itsdangerous==2.1.2
Jinja2==3.1.4
lxml==5.1.0
Markdown==3.4.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pycparser==2.21
python-dotenv==1.0.0
six==1.16.0
soupsieve==2.4
SQLAlchemy==1.4.47
SQLAlchemy-Utils==0.40.0
uWSGI==2.0.23
webencodings==0.5.1
Werkzeug==3.0.3
WTForms==3.1.2
